@@ -2,20 +2,16 @@
 name = "Algomancy"
 version = "0.2.12"
 description = "A dashboarding framework for visualizing performances of algorithms or simulations in various scenarios."
-requires-python = ">=3.11"
+requires-python = ">=3.14"
 dependencies = [
     "dash>=3.2.0",
     "dash-auth>=2.3.0",
     "dash-bootstrap-components>=2.0.3",
-    "dash-extensions>=2.0.4",
     "dash-iconify>=0.1.2",
-    "diskcache>=5.6.3",
-    "fastparquet>=2024.11.0",
     "openpyxl>=3.1.5",
     "pandas>=2.3.1",
     "prek>=0.2.1",
     "pytest>=8.4.2",
-    "python-dotenv>=1.1.1",
     "ruff>=0.13.1",
     "strenum>=0.4.15",
     "tabulate>=0.9.0",
